refactor: parsers are part of loglead package
