+++ loglead/parsers/lenma/lenma_template.py
