--- parsers/Brain/__init__.py
@@ -1,1 +1,0 @@
-from .Brain import *
